Removed docker based tests and workflow
