+++ .github/workflows/task_runner_dockerized_ws_e2e.yml
@@ -58,7 +58,7 @@ jobs:
       - name: Run Task Runner E2E tests with TLS
         id: run_tests
         run: |
-          python -m pytest -s tests/end_to_end/test_suites/task_runner_dockerized_ws_tests.py \
+          python -m pytest -s tests/end_to_end/test_suites/task_runner_tests.py \
           -m task_runner_dockerized_ws --model_name ${{ env.MODEL_NAME }} \
           --num_rounds ${{ env.NUM_ROUNDS }} --num_collaborators ${{ env.NUM_COLLABORATORS }}
           echo "Task runner end to end test run completed"
@@ -99,7 +99,7 @@ jobs:
       - name: Run Task Runner E2E tests without TLS
         id: run_tests
         run: |
-          python -m pytest -s tests/end_to_end/test_suites/task_runner_dockerized_ws_tests.py \
+          python -m pytest -s tests/end_to_end/test_suites/task_runner_tests.py \
           -m task_runner_dockerized_ws --model_name ${{ env.MODEL_NAME }} \
           --num_rounds ${{ env.NUM_ROUNDS }} --num_collaborators ${{ env.NUM_COLLABORATORS }} --disable_tls
           echo "Task runner end to end test run completed"
@@ -140,7 +140,7 @@ jobs:
       - name: Run Task Runner E2E tests without TLS
         id: run_tests
         run: |
-          python -m pytest -s tests/end_to_end/test_suites/task_runner_dockerized_ws_tests.py \
+          python -m pytest -s tests/end_to_end/test_suites/task_runner_tests.py \
           -m task_runner_dockerized_ws --model_name ${{ env.MODEL_NAME }} \
           --num_rounds ${{ env.NUM_ROUNDS }} --num_collaborators ${{ env.NUM_COLLABORATORS }} --disable_client_auth
           echo "Task runner end to end test run completed"
