+++ tests/end_to_end/models/model_owner.py
@@ -52,8 +52,7 @@ class ModelOwner():
             log.info(f"Creating workspace for model {self.model_name} at the path: {self.workspace_path}")
             error_msg = "Failed to create the workspace"
 
-            # Docker environment requires the path to be relative
-            ws_path = self.workspace_path.lstrip('/') if os.getenv("TEST_ENV") == "task_runner_docker" else self.workspace_path
+            ws_path = self.workspace_path
 
             return_code, output, error = fh.run_command(
                 f"fx workspace create --prefix {ws_path} --template {self.model_name}",
