+++ tests/end_to_end/pytest.ini
@@ -7,7 +7,6 @@ log_level = INFO
 markers =
     log_memory_usage: mark a test as a log memory usage test.
     task_runner_basic: mark a test as a task runner basic test.
-    task_runner_docker: mark a test as a task runner docker test.
-    task_runner_dockerized_ws: mark a test as a task runner dockerized ws test.
+    task_runner_dockerized_ws: mark a test as a task runner dockerized workspace test.
 asyncio_mode=auto
 asyncio_default_fixture_loop_scope="function"
